requirements.txt
Flask==1.1.2
requests==2.22.0
PyYAML==5.4.1
beautifulsoup4==4.9.3
